fixed display for textbox
made some spacing for the textbox in the server info display section, this makes the information that is displayed more readable. fixed issue where react side landing page was not being loaded and was being overwriten by server side code. removed issue in routeHandlers class that was causing that issue.